+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/com/mmxlabs/scheduler/optimiser/providers/IActualsDataProviderEditor.java
@@ -17,7 +17,7 @@ import com.mmxlabs.scheduler.optimiser.components.IPortSlot;
 public interface IActualsDataProviderEditor extends IActualsDataProvider {
 
 	void createLoadSlotActuals(ILoadOption slot, int arrivalTime, int visitDuration, long portCosts, int cargoCV, long startHeelInM3, long lngLoadVolumeInM3, long lngLoadVolumeInMMBTu,
-			int purchasePricePerMMBTu, long portBaseFuelConsumptionInMT, long ladenBaseFuelConsumptionInMT, int baseFuelPricePerMT, int charterRatePerDay, int ladenDistance, long ladenRouteCosts,
+			int purchasePricePerMMBTu, long portBaseFuelConsumptionInMT, long ladenBaseFuelConsumptionInMT, int baseFuelPricePerMT, long charterRatePerDay, int ladenDistance, long ladenRouteCosts,
 			ERouteOption route);
 
 	void createDischargeSlotActuals(IDischargeOption slot, int arrivalTime, int visitDuration, long portCosts, int cargoCV, long endHeelInM3, long lngDischargeVolumeInM3,
